calcAPC {APC}	R Documentation

Calculate the annual porcentual change

Description

This function calculates the annual porcentual change (APC) of a matrix which has the subject ID as its first column and on the rest of the columns the numeric data to calculate the APC from. Therefore, each row is a different scan ID and each column a different region.

Usage


calcAPC(fdata, f_ph)

Arguments

fdata
Data frame which has on its first column the subject ID and on the rest of the columns the data to calculate the APC from.

f_ph
Data frame, output from function load_phenotypes().

Value

Data frame

Author(s)

Lucia de Hoyos <luciadeh@gmail.com>

[Package APC version 0.1.0 Index]