load_phenotypes {APC}	R Documentation

Prepare the phenotypic data

Description

This function prepares the phenotypic data of the subjects. It inputs the phenotypic information and the column numbers of scan ID (colID), subject ID (colsID), scan timepoint (colTp) and age (colAge). It reorganises the information, to output a data frame which has the information needed to calculate the APCs for each subject and its two timepoints (scan ID and age at first and second timepoint).

Usage


load_phenotypes(file, colID = NULL, colsID = NULL, colTp = NULL, colAge = NULL)

Arguments

file
a data frame with the phenotypic information of subjects.

colID
a number defining the column where the scan ID is specified.

colsID
a number defining the column where the subject ID is specified.

colAge
a number defining the column where the age of the patient at the scan is specified

colTP
a number defining the column where the timepoint is specified. All subjects must have timepoint 1.

Value

Data frame

Author(s)

Lucia de Hoyos <luciadeh@gmail.com>

Examples


y <- load_phenotypes(x, colID = 1, colsID = 2, colTp = 3, colAge = 4)


[Package APC version 0.1.0 Index]